+++ mgl64/transform.go
@@ -13,7 +13,7 @@ import "math"
 // see HomogRotate2D
 func Rotate2D(angle float64) Mat2 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 	return Mat2{cos, sin, -sin, cos}
 }
 
@@ -25,7 +25,7 @@ func Rotate2D(angle float64) Mat2 {
 //    [0 s c ]
 func Rotate3DX(angle float64) Mat3 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 
 	return Mat3{1, 0, 0, 0, cos, sin, 0, -sin, cos}
 }
@@ -38,7 +38,7 @@ func Rotate3DX(angle float64) Mat3 {
 //    [s 0 c ]
 func Rotate3DY(angle float64) Mat3 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 
 	return Mat3{cos, 0, -sin, 0, 1, 0, sin, 0, cos}
 }
@@ -51,7 +51,7 @@ func Rotate3DY(angle float64) Mat3 {
 //    [0 0 1 ]
 func Rotate3DZ(angle float64) Mat3 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 
 	return Mat3{cos, sin, 0, -sin, cos, 0, 0, 0, 1}
 }
@@ -62,7 +62,7 @@ func Rotate3DZ(angle float64) Mat3 {
 //    [[0, 1, Ty]]
 //    [[0, 0, 1 ]]
 func Translate2D(Tx, Ty float64) Mat3 {
-	return Mat3{1, 0, 0, 0, 1, 0, float64(Tx), float64(Ty), 1}
+	return Mat3{1, 0, 0, 0, 1, 0, Tx, Ty, 1}
 }
 
 // Translate3D returns a homogeneous (4x4 for 3D-space) Translation matrix that moves a point by Tx units in the x-direction, Ty units in the y-direction,
@@ -73,20 +73,20 @@ func Translate2D(Tx, Ty float64) Mat3 {
 //    [[0, 0, 1, Tz]]
 //    [[0, 0, 0, 1 ]]
 func Translate3D(Tx, Ty, Tz float64) Mat4 {
-	return Mat4{1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, float64(Tx), float64(Ty), float64(Tz), 1}
+	return Mat4{1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, Tx, Ty, Tz, 1}
 }
 
 // Same as Rotate2D, except homogeneous (3x3 with the extra row/col being all zeroes with a one in the bottom right)
 func HomogRotate2D(angle float64) Mat3 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 	return Mat3{cos, sin, 0, -sin, cos, 0, 0, 0, 1}
 }
 
 // Same as Rotate3DX, except homogeneous (4x4 with the extra row/col being all zeroes with a one in the bottom right)
 func HomogRotate3DX(angle float64) Mat4 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 
 	return Mat4{1, 0, 0, 0, 0, cos, sin, 0, 0, -sin, cos, 0, 0, 0, 0, 1}
 }
@@ -94,14 +94,14 @@ func HomogRotate3DX(angle float64) Mat4 {
 // Same as Rotate3DY, except homogeneous (4x4 with the extra row/col being all zeroes with a one in the bottom right)
 func HomogRotate3DY(angle float64) Mat4 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 	return Mat4{cos, 0, -sin, 0, 0, 1, 0, 0, sin, 0, cos, 0, 0, 0, 0, 1}
 }
 
 // Same as Rotate3DZ, except homogeneous (4x4 with the extra row/col being all zeroes with a one in the bottom right)
 func HomogRotate3DZ(angle float64) Mat4 {
 	//angle = (angle * math.Pi) / 180.0
-	sin, cos := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	sin, cos := math.Sin(angle), math.Cos(angle)
 	return Mat4{cos, sin, 0, 0, -sin, cos, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}
 }
 
@@ -112,7 +112,7 @@ func HomogRotate3DZ(angle float64) Mat4 {
 // [[ 0     , 0     , 0     , 1 ]]
 func Scale3D(scaleX, scaleY, scaleZ float64) Mat4 {
 
-	return Mat4{float64(scaleX), 0, 0, 0, 0, float64(scaleY), 0, 0, 0, 0, float64(scaleZ), 0, 0, 0, 0, 1}
+	return Mat4{scaleX, 0, 0, 0, 0, scaleY, 0, 0, 0, 0, scaleZ, 0, 0, 0, 0, 1}
 }
 
 // Scale2D creates a homogeneous 2D scaling matrix
@@ -120,33 +120,33 @@ func Scale3D(scaleX, scaleY, scaleZ float64) Mat4 {
 // [[ 0     , scaleY, 0 ]]
 // [[ 0     , 0     , 1 ]]
 func Scale2D(scaleX, scaleY float64) Mat3 {
-	return Mat3{float64(scaleX), 0, 0, 0, float64(scaleY), 0, 0, 0, 1}
+	return Mat3{scaleX, 0, 0, 0, scaleY, 0, 0, 0, 1}
 }
 
 // ShearX2D creates a homogeneous 2D shear matrix along the X-axis
 func ShearX2D(shear float64) Mat3 {
-	return Mat3{1, 0, 0, float64(shear), 1, 0, 0, 0, 1}
+	return Mat3{1, 0, 0, shear, 1, 0, 0, 0, 1}
 }
 
 // ShearY2D creates a homogeneous 2D shear matrix along the Y-axis
 func ShearY2D(shear float64) Mat3 {
-	return Mat3{1, float64(shear), 0, 0, 1, 0, 0, 0, 1}
+	return Mat3{1, shear, 0, 0, 1, 0, 0, 0, 1}
 }
 
 // ShearX3D creates a homogeneous 3D shear matrix along the X-axis
 func ShearX3D(shearY, shearZ float64) Mat4 {
 
-	return Mat4{1, float64(shearY), float64(shearZ), 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}
+	return Mat4{1, shearY, shearZ, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}
 }
 
 // ShearY3D creates a homogeneous 3D shear matrix along the Y-axis
 func ShearY3D(shearX, shearZ float64) Mat4 {
-	return Mat4{1, 0, 0, 0, float64(shearX), 1, float64(shearZ), 0, 0, 0, 1, 0, 0, 0, 0, 1}
+	return Mat4{1, 0, 0, 0, shearX, 1, shearZ, 0, 0, 0, 1, 0, 0, 0, 0, 1}
 }
 
 // ShearZ3D creates a homogeneous 3D shear matrix along the Z-axis
 func ShearZ3D(shearX, shearY float64) Mat4 {
-	return Mat4{1, 0, 0, 0, 0, 1, 0, 0, float64(shearX), float64(shearY), 1, 0, 0, 0, 0, 1}
+	return Mat4{1, 0, 0, 0, 0, 1, 0, 0, shearX, shearY, 1, 0, 0, 0, 0, 1}
 }
 
 // HomogRotate3D creates a 3D rotation Matrix that rotates by (radian) angle about some arbitrary axis given by a Vector.
@@ -160,7 +160,7 @@ func ShearZ3D(shearX, shearY float64) Mat4 {
 //    [[ 0         , 0         , 0         , 1 ]]
 func HomogRotate3D(angle float64, axis Vec3) Mat4 {
 	x, y, z := axis[0], axis[1], axis[2]
-	s, c := float64(math.Sin(float64(angle))), float64(math.Cos(float64(angle)))
+	s, c := math.Sin(angle), math.Cos(angle)
 	k := 1 - c
 
 	return Mat4{x*x*k + c, x*y*k + z*s, x*z*k - y*s, 0, x*y*k - z*s, y*y*k + c, y*z*k + x*s, 0, x*z*k + y*s, y*z*k - x*s, z*z*k + c, 0, 0, 0, 0, 1}
@@ -168,18 +168,18 @@ func HomogRotate3D(angle float64, axis Vec3) Mat4 {
 
 // Extracts the 3d scaling from a homogeneous matrix
 func Extract3DScale(m Mat4) (x, y, z float64) {
-	return float64(math.Sqrt(float64(m[0]*m[0] + m[1]*m[1] + m[2]*m[2]))),
-		float64(math.Sqrt(float64(m[4]*m[4] + m[5]*m[5] + m[6]*m[6]))),
-		float64(math.Sqrt(float64(m[8]*m[8] + m[9]*m[9] + m[10]*m[10])))
+	return math.Sqrt(m[0]*m[0] + m[1]*m[1] + m[2]*m[2]),
+		math.Sqrt(m[4]*m[4] + m[5]*m[5] + m[6]*m[6]),
+		math.Sqrt(m[8]*m[8] + m[9]*m[9] + m[10]*m[10])
 }
 
 // Extracts the maximum scaling from a homogeneous matrix
 func ExtractMaxScale(m Mat4) float64 {
-	scaleX := float64(m[0]*m[0] + m[1]*m[1] + m[2]*m[2])
-	scaleY := float64(m[4]*m[4] + m[5]*m[5] + m[6]*m[6])
-	scaleZ := float64(m[8]*m[8] + m[9]*m[9] + m[10]*m[10])
+	scaleX := m[0]*m[0] + m[1]*m[1] + m[2]*m[2]
+	scaleY := m[4]*m[4] + m[5]*m[5] + m[6]*m[6]
+	scaleZ := m[8]*m[8] + m[9]*m[9] + m[10]*m[10]
 
-	return float64(math.Sqrt(math.Max(scaleX, math.Max(scaleY, scaleZ))))
+	return math.Sqrt(math.Max(scaleX, math.Max(scaleY, scaleZ)))
 }
 
 // Calculates the Normal of the Matrix (aka the inverse transpose)
